Move CredentialResolver to the CredentialProvider project as that makes more sense.
Fix up a number of things in the existing tests in CredentialProvider.
