--- core/src/main/resources/META-INF/services/nl.nn.adapterframework.util.AdditionalStringResolver
@@ -1,1 +1,0 @@
-nl.nn.adapterframework.util.CredentialResolver
