+++ credentialProvider/src/test/resources/StringResolver.properties
