+++ credentialProvider/src/test/resources/credentialprovider.properties
@@ -1,0 +1,2 @@
+credentialFactory.class=nl.nn.credentialprovider.PropertyFileCredentialFactory
+credentialFactory.optionalPrefix=fakePrefix:
