+++ credentialProvider/src/test/resources/credentials.properties
@@ -1,0 +1,8 @@
+
+alias1/username=username1
+alias1/password=password1
+
+alias2=passwordOnly
+
+account/username=fakeUsername
+account/password=fakePassword
